+++ lib/features/settings/application/historical_archives_workflow_panel_model_provider.g.dart
@@ -27,7 +27,7 @@ final historicalArchivesWorkflowPanelModelProvider =
 typedef HistoricalArchivesWorkflowPanelModelRef =
     AutoDisposeProviderRef<HistoricalArchivesWorkflowPanelViewModel>;
 String _$historicalArchivesWorkflowHash() =>
-    r'f94fa8724ef81ed2137463b534241debf7f5a881';
+    r'15339f2a79ea3edcdbd836fae66ba8b7e4f7ddfc';
 
 /// See also [HistoricalArchivesWorkflow].
 @ProviderFor(HistoricalArchivesWorkflow)